+++ assets/scripts/Plugin/AppHealthCheck/Interface/ICheckResult.ts
@@ -1,5 +1,0 @@
-export default interface ICheckResult
-{
-    message: string;
-    success: boolean;
-}