+++ assets/scripts/Plugin/AppHealthCheck/Interface/IHealthCheck.ts
@@ -1,6 +1,0 @@
-import ICheckResult from './ICheckResult';
-
-export default interface IHealthCheck
-{
-    check: () => Promise<ICheckResult>;
-}