+++ assets/scripts/app.ts
@@ -1,4 +1,4 @@
-import '../styles/app.scss';
+import '../styles/app.css';
 import ComponentLoader from './Core/ComponentLoader';
 import FlashHandler from './Core/FlashHandler';