+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to 
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+## Changed
+- Change capitalisation of the `DEFAULT_URI` environment variable in the installer
+
 ## [2.3.6] - 2025-05-23
 ### Added
 - Add new release workflow